style: move shared style file
