+++ wondrous-app/components/Common/Sidebar/Entity/AboutEntity/InviteButton.tsx
@@ -1,6 +1,6 @@
 import { OrgInviteLinkModal } from 'components/Common/InviteLinkModal/OrgInviteLink';
 import { PodInviteLinkModal } from 'components/Common/InviteLinkModal/podInviteLink';
-import { GreyButton } from 'components/Common/Sidebar/Common/styles';
+import { GreyButton } from 'components/Common/Sidebar/styles';
 import { useState } from 'react';
 import { useOrgBoard } from 'utils/hooks';
