+++ wondrous-app/components/Common/Sidebar/Settings/index.tsx
@@ -1,12 +1,6 @@
 import { useLazyQuery } from '@apollo/client';
 import GitHubIcon from '@mui/icons-material/GitHub';
 import { logout, useMe } from 'components/Auth/withAuth';
-import FileDownloadIcon from 'components/Icons/Sidebar/fileDownload.svg';
-import GroupIcon from 'components/Icons/Sidebar/group.svg';
-import HexagonIcon from 'components/Icons/Sidebar/hexagon.svg';
-import NotificationsIcon from 'components/Icons/Sidebar/notifications.svg';
-import PodIcon from 'components/Icons/Sidebar/pods.svg';
-import ReceiptIcon from 'components/Icons/Sidebar/receipt.svg';
 import {
   ChildrenWrapper,
   Label,
@@ -14,13 +8,19 @@ import {
   SidebarContent,
   SidebarWrapper,
   Wrapper,
-} from 'components/Common/Sidebar/Common/styles';
+} from 'components/Common/Sidebar/styles';
 import BackButton from 'components/Common/SidebarBackButton';
 import CollapseExpandButton from 'components/Common/SidebarCollapseButton';
 import Item from 'components/Common/SidebarItem';
 import ChooseEntityToCreate from 'components/CreateEntity';
 import ExitIcon from 'components/Icons/exit';
 import RolesIcon from 'components/Icons/roles';
+import FileDownloadIcon from 'components/Icons/Sidebar/fileDownload.svg';
+import GroupIcon from 'components/Icons/Sidebar/group.svg';
+import HexagonIcon from 'components/Icons/Sidebar/hexagon.svg';
+import NotificationsIcon from 'components/Icons/Sidebar/notifications.svg';
+import PodIcon from 'components/Icons/Sidebar/pods.svg';
+import ReceiptIcon from 'components/Icons/Sidebar/receipt.svg';
 import { LockIconOutline } from 'components/Icons/userpass';
 import WrenchIcon from 'components/Icons/wrench';
 import { GET_ORG_BY_ID } from 'graphql/queries/org';
